kv/list: do not rely on empty pattern match
In python < 3.7 re.split does not split a string on an empty pattern
match. So matching on the end of the string as a delimiter does not
work. E.g:
                           3.6          3.7
re.split(r'$', "foo") -> ["foo"]  ["foo", "", ""]

In order to make it work for both python versions, do not expect this
works and just consider a result of length 1 as a single last string.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>