ovs-offline: Alias ovn/ovs utilities if not available
If ovs-* / ovn-* utilities are not available locally, alias them and use
the ones shipped in the ovs-dbg container.

In order to make it work with ovn and ovs containers, we have to
sligthly refactor how the directories are organized inside the
container to make sure the default OVN_RUNDIR inside the container is
correct.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>