Remove python module
Mayor respin of the project.
The ofparse utility is now part of the official OVS python package so
there is no need to maintain it here as well. Remove it.

There is no point atm to keep a python package just for the log library.
Embed it into the utility program.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>